RM:GNL '''refactored'''
